clean code of py_layer in dygraph mode,test=develop
