+++ paddle/fluid/pybind/imperative.h
@@ -24,29 +24,6 @@ limitations under the License. */
 namespace paddle {
 namespace pybind {
 
-class Layer : public imperative::Layer {
- public:
-  using imperative::Layer::Layer;  // Inherit constructors
-
-  std::vector<imperative::VarBase*> Forward(
-      const std::vector<imperative::VarBase*>& inputs) override {
-    PYBIND11_OVERLOAD(std::vector<imperative::VarBase*>, Layer, Forward,
-                      inputs);  // NOLINT
-  }
-};
-
-class PYBIND11_HIDDEN PyOpBase : public imperative::OpBase {
- public:
-  using imperative::OpBase::OpBase;  // Inherit constructors
-
-  PyOpBase(const std::string& name) : OpBase(name) {}
-};
-
-class PyVarBase : public imperative::VarBase {
- public:
-  using imperative::VarBase::VarBase;  // Inherit constructors
-};
-
 void BindImperative(pybind11::module* m);
 
 }  // namespace pybind
